reworked file structure of project
