--- src/Components/script.js
@@ -1,5 +1,0 @@
-export default function toggleMenu() {
-    var menu = document.querySelector('#root')
-    menu.classList.toggle("sidebar-open");
-    console.log("toggle");
-}
