+++ src/layouts/RootLayout/Layout.css
@@ -9,7 +9,7 @@
     display: flex;
 }
 
-.sideBar {
+.sidebar {
     width: 0;
     background-color: #002b8e;
     transition: 0.5s cubic-bezier(0.4, 0, 0.2, 1);
@@ -19,7 +19,7 @@
     text-overflow: ellipsis;
 }
 
-.sidebar-open .sideBar {
+.sidebar-open.sidebar {
     width: 100%;
 }
