--- src/pages/Home.js
@@ -1,10 +1,0 @@
-import React from "react";
-import Container from '../Components/HomeContainer/Container';
-
-function Home() {
-    return (
-            <Container className='container'/>
-    );
-}
-
-export default Home;
