+++ src/pages/Home/Home.js
@@ -1,0 +1,10 @@
+import React from "react";
+import HomeContainer from './components/HomeContainer/HomeContainer';
+
+function Home() {
+    return (
+            <HomeContainer className='container'/>
+    );
+}
+
+export default Home;
