+++ src/utils/ToggleSidebar.js
@@ -1,0 +1,6 @@
+
+export function toggleSidebar() {
+    console.log("toggle");
+    var sidebar = document.querySelector('.sidebar')
+    sidebar.classList.toggle("sidebar-open");
+}
